Bug #1501: Remove (int) cast on Cell, Node, and Face indices to print out formatted output for regular meshes

+++ src/Core/Algorithms/Visualization/RenderField.h
@@ -1294,7 +1294,7 @@ RenderField<Fld, Loc>::render_text_nodes(FieldHandle field_handle,
     }
     else
     {
-      buffer << (int)(*iter);
+      buffer << (*iter);
     }
 
     if (use_default_material)
@@ -1487,7 +1487,7 @@ RenderField<Fld, Loc>::render_text_faces(FieldHandle field_handle,
     }
     else
     {
-      buffer << (int)(*iter);
+      buffer << (*iter);
     }
 
     if (use_default_material)
@@ -1572,7 +1572,7 @@ RenderField<Fld, Loc>::render_text_cells(FieldHandle field_handle,
     }
     else
     {
-      buffer << (int)(*iter);
+      buffer << (*iter);
     }
 
     if (use_default_material)
